# docs/openapi/spec/membershipfees.yaml
...
openapi: 3.1.0
info:
  title: Klabis API — Membership fees module
  description: The Membership fees slice of the Klabis API. See klabis.yaml for the full API.
  version: 0.1.0
servers:
  - url: https://localhost:8443
    description: Local development server
  - url: https://api.klabis.com
    description: Production server
paths:
  /api/membership-fee-tiers:
    get:
      operationId: listTiers
      tags: [MembershipFeeTiers]
      summary: List all membership fee tiers
      description: Returns all membership fee tiers in the catalog.
      security:
        - KlabisAuth: [MEMBERS]
      # No x-klabis-authority: readable by any authenticated caller. The activeCampaign/
      # pastCampaigns links and the createTier/publishYear affordances are only added for callers
      # with MEMBERS:MANAGE, but that is decided inline in the controller (isAdmin()) rather than
      # gating the whole operation.
      responses:
        '200':
          description: List of membership fee tiers retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MembershipFeeTierSummaryResponseList'
            application/prs.hal-forms+json:
              schema:
                $ref: '#/components/schemas/CollectionModelEntityModelMembershipFeeTierSummaryResponse'
          x-hal-links:
            self:
              description: This collection
            activeCampaign:
              operation: getPublication
              description: Present only for callers with MEMBERS:MANAGE, when a campaign is active
            pastCampaigns:
              operation: listPublications
              description: Present only for callers with MEMBERS:MANAGE
          x-hal-templates:
            createTier:
              operation: createTier
              description: Present only for callers with MEMBERS:MANAGE
            publishYear:
              operation: publishYear
              description: Present only for callers with MEMBERS:MANAGE
        '400':
          $ref: './_shared/responses.yaml#/components/responses/BadRequest'
        '401':
          $ref: './_shared/responses.yaml#/components/responses/Unauthorized'
        '403':
          $ref: './_shared/responses.yaml#/components/responses/Forbidden'
        '404':
          $ref: './_shared/responses.yaml#/components/responses/NotFound'
        '409':
          $ref: './_shared/responses.yaml#/components/responses/Conflict'
        '422':
          $ref: './_shared/responses.yaml#/components/responses/UnprocessableEntity'

    post:
      operationId: createTier
      tags: [MembershipFeeTiers]
      summary: Create a membership fee tier
      description: Creates a new membership fee tier in the catalog.
      security:
        - KlabisAuth: [MEMBERS]
      x-klabis-authority: MEMBERS_MANAGE
      requestBody:
        required: true
        x-codegen-request-body-name: request
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMembershipFeeTierRequest'
      responses:
        '201':
          description: Membership fee tier created
          headers:
            Location:
              description: URI of the created tier
              schema:
                type: string
                format: uri
          content:
            application/prs.hal-forms+json: {}
        '400':
          $ref: './_shared/responses.yaml#/components/responses/BadRequest'
        '401':
          $ref: './_shared/responses.yaml#/components/responses/Unauthorized'
        '403':
          $ref: './_shared/responses.yaml#/components/responses/Forbidden'
        '404':
          $ref: './_shared/responses.yaml#/components/responses/NotFound'
        '409':
          $ref: './_shared/responses.yaml#/components/responses/Conflict'
        '422':
          $ref: './_shared/responses.yaml#/components/responses/UnprocessableEntity'
...
  /api/fee-selection-campaigns:
    get:
      operationId: listPublications
      tags: [FeeSelectionCampaigns]
      summary: List fee year publications, optionally filtered by status
      description: |
        Lists fee selection campaigns (one per calendar year). status=closed returns only past
        campaigns; omitted returns all.
      security:
        - KlabisAuth: [MEMBERS]
      # No x-klabis-authority: readable by any authenticated caller. publishYear affordance still
      # requires MEMBERS:MANAGE via klabisAfford's own authorization check.
      parameters:
        - name: status
          in: query
          description: 'Filter by status: ''closed'' returns only past campaigns'
          required: false
          schema:
            # Plain string, not enum: an inline `enum:` on a property/parameter that IS generated
            # into Java (request DTOs, query params) produces a reference to a synthetic
            # "<Schema><Property>" (or named-schema) enum class, but this generator version emits
            # an EMPTY body for a bare top-level enum schema used only as a query-param type — the
            # file compiles to nothing and every reference fails "cannot find symbol". No existing
            # domain enum exists to map onto either (see events.yaml's EventStatus precedent, which
            # works because it maps onto a real domain enum). Plain string preserves the actual
            # wire format; CampaignStatusFilter.valueOf() in the controller already validates it.
            type: string
      responses:
        '200':
          description: List of fee year publications retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeeSelectionCampaignResponseList'
            application/prs.hal-forms+json:
              schema:
                $ref: '#/components/schemas/CollectionModelEntityModelFeeSelectionCampaignResponse'
          x-hal-links:
            self:
              description: This collection, with the current status filter
          x-hal-templates:
            publishYear:
              operation: publishYear
              description: Present only for callers with MEMBERS:MANAGE
        '400':
          $ref: './_shared/responses.yaml#/components/responses/BadRequest'
        '401':
          $ref: './_shared/responses.yaml#/components/responses/Unauthorized'
        '403':
          $ref: './_shared/responses.yaml#/components/responses/Forbidden'
        '404':
          $ref: './_shared/responses.yaml#/components/responses/NotFound'
        '409':
          $ref: './_shared/responses.yaml#/components/responses/Conflict'
        '422':
          $ref: './_shared/responses.yaml#/components/responses/UnprocessableEntity'

    post:
      operationId: publishYear
      tags: [FeeSelectionCampaigns]
      summary: Publish fee levels for a calendar year
      description: |
        Opens a fee selection campaign for a calendar year, publishing a snapshot of the given
        membership fee tiers as fee groups members can choose from.
      security:
        - KlabisAuth: [MEMBERS]
      x-klabis-authority: MEMBERS_MANAGE
      requestBody:
        required: true
        x-codegen-request-body-name: request
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PublishYearRequest'
      responses:
        '201':
          description: Fee year publication created
          headers:
            Location:
              description: URI of the created publication
              schema:
                type: string
                format: uri
          content:
            application/prs.hal-forms+json: {}
        '400':
          $ref: './_shared/responses.yaml#/components/responses/BadRequest'
        '401':
          $ref: './_shared/responses.yaml#/components/responses/Unauthorized'
        '403':
          $ref: './_shared/responses.yaml#/components/responses/Forbidden'
        '404':
          $ref: './_shared/responses.yaml#/components/responses/NotFound'
        '409':
          $ref: './_shared/responses.yaml#/components/responses/Conflict'
        '422':
          $ref: './_shared/responses.yaml#/components/responses/UnprocessableEntity'
  /api/fee-selection-campaigns/{id}:
    get:
      operationId: getPublication
      tags: [FeeSelectionCampaigns]
      summary: Get fee year publication details
      description: Returns a single fee selection campaign by ID.
      security:
        - KlabisAuth: [MEMBERS]
      # No x-klabis-authority: readable by any authenticated caller. changeDeadline/closeCampaign
      # affordances still require MEMBERS:MANAGE via klabisAfford's own authorization check, and are
      # additionally state-gated (not closed / not processed) in the postprocessor.
      parameters:
        - $ref: '#/components/parameters/CampaignIdParam'
      responses:
        '200':
          description: Fee year publication found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeeSelectionCampaignResponse'
            application/prs.hal-forms+json:
              schema:
                $ref: '#/components/schemas/EntityModelFeeSelectionCampaignResponse'
          x-hal-links:
            self:
              description: This publication
            collection:
              operation: listPublications
              description: Back to the publication list
            levels:
              operation: listGroupsForYear
              description: Fee groups published for this campaign's year
          x-hal-templates:
            default:
              operation: changeDeadline
              description: Present for callers with MEMBERS:MANAGE, while the campaign is not closed
            closeCampaign:
              operation: closeCampaign
              description: |
                Present for callers with MEMBERS:MANAGE, while the campaign is not closed and not
                yet processed
        '400':
          $ref: './_shared/responses.yaml#/components/responses/BadRequest'
        '401':
          $ref: './_shared/responses.yaml#/components/responses/Unauthorized'
        '403':
          $ref: './_shared/responses.yaml#/components/responses/Forbidden'
        '404':
          $ref: './_shared/responses.yaml#/components/responses/NotFound'
        '409':
          $ref: './_shared/responses.yaml#/components/responses/Conflict'
        '422':
          $ref: './_shared/responses.yaml#/components/responses/UnprocessableEntity'
  /api/fee-selection-campaigns/{id}/deadline:
    patch:
      operationId: changeDeadline
      tags: [FeeSelectionCampaigns]
      summary: Change voting deadline of an active campaign
      description: Updates the voting deadline of a campaign that is not yet closed.
      security:
        - KlabisAuth: [MEMBERS]
      x-klabis-authority: MEMBERS_MANAGE
      parameters:
        - $ref: '#/components/parameters/CampaignIdParam'
      requestBody:
        required: true
        x-codegen-request-body-name: request
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangeDeadlineRequest'
      responses:
        '200':
          description: Voting deadline updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeeSelectionCampaignResponse'
            application/prs.hal-forms+json:
              schema:
                $ref: '#/components/schemas/EntityModelFeeSelectionCampaignResponse'
          x-hal-links:
            self:
              description: This publication
        '400':
          $ref: './_shared/responses.yaml#/components/responses/BadRequest'
        '401':
          $ref: './_shared/responses.yaml#/components/responses/Unauthorized'
        '403':
          $ref: './_shared/responses.yaml#/components/responses/Forbidden'
        '404':
          $ref: './_shared/responses.yaml#/components/responses/NotFound'
        '409':
          $ref: './_shared/responses.yaml#/components/responses/Conflict'
        '422':
          $ref: './_shared/responses.yaml#/components/responses/UnprocessableEntity'
  /api/fee-selection-campaigns/{id}/close:
    post:
      operationId: closeCampaign
      tags: [FeeSelectionCampaigns]
      summary: Manually close a campaign
      description: |
        Processes a campaign immediately, regardless of its voting deadline — finalising member
        assignments to fee groups for the campaign's year.
      security:
        - KlabisAuth: [MEMBERS]
      x-klabis-authority: MEMBERS_MANAGE
      parameters:
        - $ref: '#/components/parameters/CampaignIdParam'
      responses:
        '204':
          description: Campaign closed
          content:
            application/prs.hal-forms+json: {}
        '400':
          $ref: './_shared/responses.yaml#/components/responses/BadRequest'
        '401':
          $ref: './_shared/responses.yaml#/components/responses/Unauthorized'
        '403':
          $ref: './_shared/responses.yaml#/components/responses/Forbidden'
        '404':
          $ref: './_shared/responses.yaml#/components/responses/NotFound'
        '409':
          $ref: './_shared/responses.yaml#/components/responses/Conflict'
        '422':
          $ref: './_shared/responses.yaml#/components/responses/UnprocessableEntity'
  /api/fee-selection-campaigns/{year}/levels:
    get:
      operationId: listGroupsForYear
      tags: [FeeSelectionCampaigns]
      summary: List published fee groups for a given year
      description: Returns the fee groups published as part of the campaign for the given year.
      security:
        - KlabisAuth: [MEMBERS]
      # No x-klabis-authority: readable by any authenticated caller.
      parameters:
        - $ref: '#/components/parameters/CampaignYearParam'
      responses:
        '200':
          description: List of fee groups retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MembershipFeeGroupResponseList'
            application/prs.hal-forms+json:
              schema:
                $ref: '#/components/schemas/CollectionModelEntityModelMembershipFeeGroupResponse'
          x-hal-links:
            self:
              description: This collection, for the given year
        '400':
          $ref: './_shared/responses.yaml#/components/responses/BadRequest'
        '401':
          $ref: './_shared/responses.yaml#/components/responses/Unauthorized'
        '403':
          $ref: './_shared/responses.yaml#/components/responses/Forbidden'
        '404':
          $ref: './_shared/responses.yaml#/components/responses/NotFound'
        '409':
          $ref: './_shared/responses.yaml#/components/responses/Conflict'
        '422':
          $ref: './_shared/responses.yaml#/components/responses/UnprocessableEntity'
...
components:
  securitySchemes:
    KlabisAuth:
      type: oauth2
      description: OAuth2 authentication with Klabis Authorization Server
      flows:
        authorizationCode:
          authorizationUrl: /oauth2/authorize
          tokenUrl: /oauth2/token
          scopes:
            openid: authenticate
            profile: display user profile information
            MEMBERS: Members
            EVENTS: Events
  parameters:
...
    CampaignIdParam:
      name: id
      in: path
      description: Fee selection campaign UUID
      required: true
      schema:
        type: string
        format: uuid
    CampaignYearParam:
      name: year
      in: path
      description: Calendar year
      required: true
      schema:
        type: integer
        format: int32
...
  schemas:
    MembershipFeeTierSummaryResponseList:
      type: array
      items:
        $ref: '#/components/schemas/MembershipFeeTierSummaryResponse'

    # Envelope schemas — never generated into Java; see klabis-api-spec skill:
    # "Payload and envelope are separate schemas".
    CollectionModelEntityModelMembershipFeeTierSummaryResponse:
      type: object
      properties:
        _embedded:
          type: object
          properties:
            membershipFeeTierSummaryResponseList:
              type: array
              items:
                $ref: '#/components/schemas/EntityModelMembershipFeeTierSummaryResponse'
        _links:
          $ref: './_shared/hal.yaml#/components/schemas/Links'
        _templates:
          $ref: './_shared/hal.yaml#/components/schemas/HalFormsTemplates'
    EntityModelMembershipFeeTierSummaryResponse:
      allOf:
        - $ref: '#/components/schemas/MembershipFeeTierSummaryResponse'
        - type: object
          properties:
            _links:
              $ref: './_shared/hal.yaml#/components/schemas/Links'
    MembershipFeeTierSummaryResponse:
      type: object
      description: Membership fee tier summary for list views.
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        yearlyFeeAmount:
          type: number
        yearlyFeeCurrency:
          type: string
        ruleCount:
          type: integer
          format: int32
...
    CreateMembershipFeeTierRequest:
      type: object
      required: [name, yearlyFeeAmount]
      properties:
        name:
          type: string
          x-klabis-not-blank: true
        yearlyFeeAmount:
          type: number
          exclusiveMinimum: 0
        yearlyFeeCurrency:
          type: string
          description: Defaults to CZK when omitted.
...
    PaymentRuleResponse:
      type: object
      description: |
        A single payment rule. percentage is present when ruleType is PERCENTAGE; fixedAmount/
        fixedCurrency are present when ruleType is FIXED_AMOUNT.
      properties:
        eventTypeId:
          type: string
          format: uuid
        rankingShortName:
          type: string
        ruleType:
          type: string
          enum: [PERCENTAGE, FIXED_AMOUNT]
        percentage:
          type: integer
          format: int32
        fixedAmount:
          type: number
        fixedCurrency:
          type: string
...
    FeeSelectionCampaignResponseList:
      type: array
      items:
        $ref: '#/components/schemas/FeeSelectionCampaignResponse'
    CollectionModelEntityModelFeeSelectionCampaignResponse:
      type: object
      properties:
        _embedded:
          type: object
          properties:
            feeSelectionCampaignResponseList:
              type: array
              items:
                $ref: '#/components/schemas/EntityModelFeeSelectionCampaignResponse'
        _links:
          $ref: './_shared/hal.yaml#/components/schemas/Links'
        _templates:
          $ref: './_shared/hal.yaml#/components/schemas/HalFormsTemplates'
    EntityModelFeeSelectionCampaignResponse:
      allOf:
        - $ref: '#/components/schemas/FeeSelectionCampaignResponse'
        - type: object
          properties:
            _links:
              $ref: './_shared/hal.yaml#/components/schemas/Links'
            _templates:
              $ref: './_shared/hal.yaml#/components/schemas/HalFormsTemplates'
    FeeSelectionCampaignResponse:
      type: object
      description: A fee selection campaign for one calendar year.
      properties:
        id:
          type: string
          format: uuid
        year:
          type: integer
          format: int32
        votingDeadline:
          type: string
          format: date
        deadlineProcessedAt:
          type: string
          format: date-time
    PublishYearRequest:
      type: object
      required: [year, votingDeadline, levelIds]
      properties:
        year:
          type: integer
          format: int32
          exclusiveMinimum: 0
        votingDeadline:
          type: string
          format: date
        levelIds:
          type: array
          minItems: 1
          items:
            type: string
            format: uuid
    ChangeDeadlineRequest:
      type: object
      required: [votingDeadline]
      properties:
        votingDeadline:
          type: string
          format: date
    MembershipFeeGroupResponseList:
      type: array
      items:
        $ref: '#/components/schemas/MembershipFeeGroupResponse'
    CollectionModelEntityModelMembershipFeeGroupResponse:
      type: object
      properties:
        _embedded:
          type: object
          properties:
            membershipFeeGroupResponseList:
              type: array
              items:
                $ref: '#/components/schemas/EntityModelMembershipFeeGroupResponse'
        _links:
          $ref: './_shared/hal.yaml#/components/schemas/Links'
    EntityModelMembershipFeeGroupResponse:
      allOf:
        - $ref: '#/components/schemas/MembershipFeeGroupResponse'
        - type: object
          properties:
            _links:
              $ref: './_shared/hal.yaml#/components/schemas/Links'

    # The full wire shape of getFeeGroup: MembershipFeeGroupResponse's properties plus an
    # _embedded.members (per @Relation(collectionRelation = "members") on MemberInGroupResponse).
    # schemaMappings points at MembershipFeeGroupResponse, so the generated method returns the
    # payload alone — the _embedded block is declared by the controller
    # (HalResponseContext.embed) and assembled by HalResponseBodyAdvice, like _links and _templates.
...
    MembershipFeeGroupResponse:
      type: object
      description: A published fee group — a snapshot taken from a tier for one campaign year.
      properties:
        id:
          type: string
          format: uuid
        sourceLevelId:
          type: string
          format: uuid
        name:
          type: string
        year:
          type: integer
          format: int32
        yearlyFeeAmount:
          type: number
        yearlyFeeCurrency:
          type: string
        status:
          type: string
          enum: [EDITABLE, FROZEN]
        memberCount:
          type: integer
          format: int32
        rulesSnapshot:
          type: array
          items:
            $ref: '#/components/schemas/PaymentRuleResponse'
...